// src/models/AdvertModel.ts
export interface AdvertModel {
    id: number
    userId: string
    cityId: number
    cityName: string
    areaName: string
    areaId:number
    categoryId: number
    categoryName: string
    date: string
    title: string
    description: string
    isNew: boolean
    isVip: boolean
    isContractPrice: boolean
    price: number
    firstImage: string
    isFavorite: boolean
    contactEmail: string
    phoneNumber: string
    contactPersone: string
}